Added variables.ipynb tutorial on Python variables
This commit adds the a Jupyter Notebook (02_variables.ipynb) that serves as a tutorial on Python variables. The notebook covers key concepts such as variable. It includes practical examples and code snippets to help learners understand how variables work in Python.

This tutorial is designed for beginners and can be useful for anyone looking to strengthen their foundational Python skills.